fix: bugfix & renamed package
